tests/wpt/web-platform-tests/fullscreen/rendering/ua-style-iframe-manual.html

<!DOCTYPE html>
<title>User-agent levels style sheet defaults for iframe</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<style>
iframe {
  border: 1px solid blue;
  padding: 1px;
  /* transform is also tested because of https://crbug.com/662393 */
  transform: scale(0.5);
}
</style>
<div id="log"></div>
<div id="ancestor"><iframe></iframe></div>
<script>
async_test(t => {
  const ancestor = document.getElementById('ancestor');
  const iframe = ancestor.firstChild;

  const initialStyle = getComputedStyle(iframe);
  assert_equals(initialStyle.border, '1px solid rgb(0, 0, 255)', 'initial border style');
  assert_equals(initialStyle.padding, '1px', 'initial padding style');
  assert_equals(initialStyle.transform, 'matrix(0.5, 0, 0, 0.5, 0, 0)', 'initial transform style');

  trusted_request(t, iframe);

  document.addEventListener('fullscreenchange', t.step_func_done(() => {
    const fullscreenStyle = getComputedStyle(iframe);
    assert_equals(fullscreenStyle.border, '0px none rgb(0, 0, 0)', 'fullscreen border style');
    assert_equals(fullscreenStyle.padding, '0px', 'fullscreen padding style');
    assert_equals(fullscreenStyle.transform, 'none', 'fullscreen transform style');
  }));
});
</script>
